Refactor token amount formatting and validation logic
Moved and improved the minimal token amount formatting function to utils/format.ts for reuse. Updated RequestLiquidityDialog to use clearer validation variables for form submission. Fixed rounding mode constant in useRequestLiquidity. Optimized getDefaultUsdcTokenId for a fast path symbol match.
