+++ utils/format.ts
@@ -6,3 +6,25 @@ export function parseNumber(input: string | number | null | undefined): number {
   return Number.isNaN(n) ? NaN : n;
 }
 
+/**
+ * Format a minimal-unit token amount into a human-readable decimal string.
+ *
+ * Example: minimal="0012300", decimals=4 → "12.3"
+ *
+ * Rules:
+ * - Fast-path zero: returns "0" for inputs like "0", "00", …
+ * - Inserts the decimal point according to `decimals`
+ * - Trims superfluous leading zeros, trailing zeros, and trailing decimal point
+ */
+export function formatMinimalTokenAmount(minimal: string, decimals: number): string {
+  if (/^0+$/.test(minimal)) return "0";
+  const s = minimal.replace(/^0+(?=\d)/, "");
+  const d = Math.max(0, decimals);
+  const pad = s.length <= d ? "0".repeat(d - s.length + 1) + s : s;
+  const i = pad.length - d;
+  const withDot = d === 0 ? pad : `${pad.slice(0, i)}.${pad.slice(i)}`;
+  return withDot
+    .replace(/^0+(\d)/, "$1")
+    .replace(/\.0+$/, "")
+    .replace(/(\.\d*?)0+$/, "$1");
+}
